+++ src/test/java/com/github/camilormz/finalreality/model/character/EnemyTestBK.java
@@ -1,9 +1,0 @@
-package com.github.camilormz.finalreality.model.character;
-
-// TODO: This test IS totally broken and must be re-done ASAP
-
-class EnemyTestBK extends AbstractCharacterTestBK {
-
-  private static final String ENEMY_NAME = "Goblin";
-
-}